Added a mixin to the condition and encounter filters to remove duplicate code and fixed the encounter filter
